Update installation procedure and guide

+++ bootstrap-apache.sh
@@ -26,11 +26,15 @@ echo "postfix postfix/mailname string precise32" | debconf-set-selections
 apt-get -y install postfix
 
 # use sample configs for quick startup
-cp /kisakone/config.php.sample /kisakone/config.php
 cp /kisakone/config_site.php.sample /kisakone/config_site.php
 
 # restore backups (expects backup file to contain proper "use database;")
-for FILE in $(ls -1 /vagrant/*.sql.backup); do
-  echo "Running SQL from: $FILE"
-  mysql -u root --password=pass < $FILE
-done
+if [ -e "/vagrant/kisakone.sql.backup" ]; then
+  echo "Restoring backup SQL from: kisakone.sql.backup"
+  mysql -u root --password=pass < /vagrant/kisakone.sql.backup
+  cp /kisakone/config.php.sample /kisakone/config.php
+  echo "Done! Kisakone restored from backup at http://127.0.0.1/"
+else
+  echo "Done! Kisakone is waiting for installation at http://127.0.0.1/doc/install/install.php"
+fi
+
